% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.frame.R
\docType{data}
\name{as.data.frame}
\alias{as.data.frame}
\alias{as.data.frame,ERGMarker-method}
\alias{as.data.frame,ERGChannel-method}
\alias{as.data.frame,ERGStep-method}
\title{as.data.frame for ERGProtocol, ERGStep, ERGChannel and ERGMarker}
\format{
An object of class \code{character} of length 1.
}
\usage{
\S4method{as.data.frame}{ERGMarker}(x, row.names = NULL, optional = FALSE, ...)

\S4method{as.data.frame}{ERGChannel}(x, row.names = NULL, optional = FALSE, ...)

\S4method{as.data.frame}{ERGStep}(x, row.names = NULL, optional = FALSE, ...)

as.data.frame
}
\arguments{
\item{x}{An \link{ERGProtocol}, ERGStep, ERGChannel or ERGMarker object.}

\item{...}{currently unused.}
}
\value{
A data frame representing the \link{ERGProtocol}, ERGStep, ERGChannel or ERGMarker object in long format.
}
\description{
Converts an \link{ERGExam}, \link{ERGProtocol}, ERGStep, ERGChannel or ERGMarker object to a data frame format.
Note that the \code{as.data.frame} method for \link{ERGExam} objects is inherited from the \link[EPhysData]{EPhysData-package} package. See: \link[EPhysData:as.data.frame]{EPhysData::as.data.frame}
}
\section{Functions}{
\itemize{
\item \code{as.data.frame(ERGMarker)}: Method for ERGMarker

\item \code{as.data.frame(ERGChannel)}: Method for ERGChannel

\item \code{as.data.frame(ERGStep)}: Method for ERGStep

\item \code{as.data.frame}: Method for ERGProtocol

}}
\examples{
# Create an example EPhysData object
new_data <- makeExampleEPhysData()
# Convert EPhysData to a data frame
new_data <- as.data.frame(new_data)
new_data

# Create an example EPhysSet object
new_data <- makeExampleEPhysSet()
new_data <- as.data.frame(new_data)

}
\keyword{datasets}
