% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Changelog.R
\name{Changelog}
\alias{Changelog}
\title{Display Changelog for an ERGExam Object}
\usage{
Changelog(X, return = F)
}
\arguments{
\item{X}{An object of class \code{ERGExam}.}

\item{return}{should the changelog be returned as a variable?}
}
\value{
Invisibly returns the original \code{ERGExam} object.
}
\description{
This method returns the formatted changelog for an ERGExam object. It uses
the cli package to format the output. If no changelog entries are present,
an informational message is displayed.
}
\examples{
\dontrun{
  # Assuming exam is an instance of ERGExam with a changelog slot
  Changelog(exam)
}

}
