% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Checks.R
\name{CheckAvgFxSet}
\alias{CheckAvgFxSet}
\title{Check if all recordings in an ERGExam object have a valid averaging function set.}
\usage{
CheckAvgFxSet(X, fatal = TRUE)
}
\arguments{
\item{X}{An ERGExam object.}

\item{fatal}{Logical, indicating whether to stop with an error message if not all recordings have a valid averaging function.}
}
\value{
Logical, TRUE if all recordings have a valid averaging function, FALSE otherwise.
}
\description{
Check if all recordings in an ERGExam object have a valid averaging function set.
}
\examples{
CheckAvgFxSet(myERGExamObject)

}
\seealso{
\code{\link{AverageFunction<-}}, \code{\link{SetStandardFunctions}}
}
