% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clear.R
\name{Clear}
\alias{Clear}
\title{Clear Computed Slots from an ERGExam object}
\value{
An ERGExam object with the Results slot cleared.
}
\description{
Clears the Averaged and Measurements slots in an ERGExam object, respectiveley, by removing any existing data.
}
\examples{
# Clear the Results slot of an ERGExam object
cleared_exam <- clearResults(exam)
}
