% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasurementMethots.R
\name{ConvertMeasurementsToAbsolute}
\alias{ConvertMeasurementsToAbsolute}
\title{Measurements-Methods Converts the output of the Measurements method from relative to absolute amplitudes}
\usage{
ConvertMeasurementsToAbsolute(data)
}
\description{
Measurements-Methods Converts the output of the Measurements method from relative to absolute amplitudes
}
\keyword{Internal}
