% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0ERGElectrode-class.R
\docType{class}
\name{ERGElectrode-class}
\alias{ERGElectrode-class}
\title{The ERGElectrode Class}
\description{
An S4 class representing an electrode entry.
}
\section{Slots}{

\describe{
\item{\code{Name}}{Character. The identifier of the electrode (e.g. "Ground").}

\item{\code{Type}}{Character. The type of the electrode (e.g. "Platinum").}

\item{\code{Position}}{Character. The position of the electrode ("E.g. Tail").}

\item{\code{Impedance}}{Numeric of class units. The impedance of the electrode. Unit must be convertible to Ohm.}
}}

\examples{
e1 <- newERGElectrode("Ground", "Platinum", "Tail", as_units(100, "Ohm"))
e1
}
