% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0ERGExam.R
\docType{class}
\name{ERGExam-class}
\alias{ERGExam-class}
\alias{ERGExam}
\title{An S4 class storing an ERG Exam}
\description{
This object is designed to store data and metadata from a single ERG experiment/exam
}
\section{Slots}{

\describe{
\item{\code{Measurements}}{A data.frame containing results from measurements performed on ERG traces. Each row represents one measurement}

\item{\code{Channels}}{A character vector containing the names of the recorded channels}

\item{\code{Stimulus}}{A data.frame containing information on the stimuli employed. Each row represents one step.}

\item{\code{ProtocolName,Version,Filename,RecMode,Patient,Gender,Group,Investigator}}{Character vectors.}

\item{\code{ExamDate,DOB,Imported}}{POSIXct/Date containing time stamps of the time/date the examination was performed, the study subject was born and data were imported, resp.}

\item{\code{Steps_Timetrace,Steps_RAW,Steps_AVG,Steps_Excl}}{Reserved for storing raw and averaged traced data. Currently unsupported.}
}}

\seealso{
\linkS4class{ERGExam}, \linkS4class{ImportEpsionMeasures}
}
