% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1ERGExam-class.R
\docType{class}
\name{ERGExam}
\alias{ERGExam}
\alias{ERGExam-class}
\title{ERGExam Class}
\description{
A class representing an ERG (Electroretinogram) exam with associated data and attributes. This class extends the \link[EPhysData:EPhysSet]{EPhysData::EPhysSet} object and all methods valid for \link[EPhysData:EphysSet]{EPhysData::EphysSet} can also be applied to \code{ERGExam} objects.
}
\section{Slots}{

\describe{
\item{\code{Data}}{Data A list of \link{EPhysData} objects. Each item containing a recording in respsonse to a particular Stimulus ("Step"), from a particular eye and data Channel (e.g. ERG or OP)-}

\item{\code{Metadata}}{A data frame containing metadata information associated with the data, each row corresponds to one item in \code{data}.
\describe{
\item{Step}{A character vector containing the step name. A step describes data recorded in response to the same type of stimulus.}
\item{Eye}{A character vector. Possible values "RE" (right eye) and "LE" (left eye).}
\item{Channel}{A character vector containing the channel name. This can be "ERG", "VEP" or "OP" for instance.}
\item{Result}{A numeric vector containing the indices of individual results contained in an ERG exam. E.g., if a recording to one identical stimulus is performed twice, these would be distinguished by different indices in the Result column. Warning: This is currently experimental}
}}

\item{\code{Stimulus}}{A data frame containing stimulus information.
\describe{
\item{Step}{A character vector containing the step the given stimulus is associated with.}
\item{Description}{A character vector describing the stimulus.}
\item{Intensity}{An integer vector representing the intensity of the stimulus.}
\item{Background}{A character vector describing the adaptation state of the retina for that stimulus (DA or LA).}
\item{Type}{A character vector describing the type of the stimulus (e.g. Flash or Flicker).}
}}

\item{\code{Averaged}}{TRUE if the object contains averaged data, FALES indicates object contains raw traces.}

\item{\code{Measurements}}{An object of class \linkS4class{ERGMeasurements}}

\item{\code{ExamInfo}}{A list containing exam-related information.
\describe{
\item{ProtocolName}{A character vector indicating the name of the protocol.}
\item{Version}{Optional: A character vector indicating the version of the protocol.}
\item{ExamDate}{A \code{POSIXct} The date of the exam.}
\item{Filename}{Optional: The filename associated where exam raw data have been imported from.}
\item{RecMode}{Optional: A character vector indicating the recording mode.}
\item{Investigator}{Optional: A character vector indicating the name of the investigator conducting the exam.}
}}

\item{\code{SubjectInfo}}{A list containing subject-related information.
\describe{
\item{Subject}{A character vector indicating the name of the subject}
\item{DOB}{A \code{Date} object indicating the date of birth of the subject}
\item{Gender}{Optional: A character vector indicating the gender of the subject}
\item{Group}{Optional: A character vector indicating the study group to which the subject belongs.}
}}

\item{\code{Imported}}{A \code{POSIXct} timestamp indicating when the object was imported.}
}}

\seealso{
\link[EPhysData:EPhysData]{EPhysData::EPhysData-package} \link[EPhysData:EPhysData]{EPhysData::EPhysData-class} \link[EPhysData:EPhysSet]{EPhysData::EPhysSet-class}
}
