% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get-methods.R
\name{Get}
\alias{Get}
\alias{GetStimulusNames}
\alias{GetMarkerNames}
\alias{GetStimulusTable}
\alias{GetProtocolName}
\alias{GetGroupName}
\alias{GetPatientName}
\alias{GetExamDate}
\alias{GetDOB}
\title{Accession methods for metadata from ERGExam objects}
\usage{
GetChannelNames(X)

GetStimulusNames(X)

GetMarkerNames(X)

GetStimulusTable(X)

GetProtocolName(X)

GetGroupName(X)

GetPatientName(X)

GetExamDate(X)

GetDOB(X)

GetStimulusNames(X)
}
\arguments{
\item{X}{A \linkS4class{ERGExam} object}
}
\value{
A character vector. For GetStimulusTable a data.frame.
}
\description{
`r lifecycle::badge("stable")` \cr
These methods are used to access metadata information from \linkS4class{ERGExam} objects.
}
\details{
These methods can be used to access metadata information stored in \linkS4class{ERGExam} objects. \cr \cr
}
\section{Functions}{
\itemize{
\item \code{GetStimulusNames}: Returns the stimulus names.

\item \code{GetMarkerNames}: Returns the measurement parameter names (e.g: 'a','B','N1','P1').

\item \code{GetStimulusTable}: Returns the stimulus table

\item \code{GetProtocolName}: Returns the recording protocol name.

\item \code{GetGroupName}: Returns the group name.

\item \code{GetPatientName}: Returns the subject name/ID.

\item \code{GetExamDate}: Returns the exam date.

\item \code{GetDOB}: Returns the date of birth.

\item \code{GetStimulusNames}: Returns the stimulus names.
}}

