% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetData.R
\name{GetData}
\alias{GetData}
\title{Get measurements data from an ERGExam object}
\usage{
GetData(
  X,
  Channels = GetChannelNames(X),
  Stimuli = GetStimulusNames(X),
  Repeats = "ALL",
  Markers = GetMarkerNames(X),
  Measures = c("Voltage", "Time"),
  Verbose = T
)
}
\arguments{
\item{Channels, Stimuli, Markers}{List of Channels/Stimuli or measurement parameters (e.g. a-wave) to get. By default will get all.}

\item{Repeats}{Measurement repeats to get. Can be "ALL", to get all repeats, a numeric vector of length >=1 containing the repeats to keep, or a function of the form \code{function(x){x==max(x)}} indicating which repeats to keep.}

\item{Measures}{Measures to get. A character vector. Can be \code{"Voltage"}, \code{"Time"} or both (default).}

\item{Verbose}{Show warning/info messages.}
}
\value{
An \linkS4class{ERGExam} object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}} \cr
This function subsets \linkS4class{ERGExam} objects by \var{Channels}, \var{Stimuli}, \var{Repeats}, measurement \var{Markers} or \var{Measures}.
}
\examples{
\dontrun{
# import a *.csv files exported from Diagnosys' Epsion
ERG_Experiment<-ImportEpsionMeasures("test.csv")
# get all measurements with the latest (highest number) repeat
df<-GetData(ERG_Experiment,Repeats = function(x){x==max(x)},verbose=T)
}
}
