% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMeasurements.R
\name{GetMeasurements}
\alias{GetMeasurements}
\alias{GetAverageTraces}
\title{Get data from an ERGExam object}
\usage{
GetMeasurements(
  X,
  Channels = GetChannelNames(X),
  Stimuli = GetStimulusNames(X),
  Repeats = "ALL",
  Markers = GetMarkerNames(X),
  Measures = c("Voltage", "Time"),
  Verbose = T
)

GetAverageTraces(
  X,
  Channels = GetChannelNames(X),
  Stimuli = GetStimulusNames(X),
  Repeats = "ALL",
  Verbose = T
)
}
\arguments{
\item{Channels, Stimuli, Markers}{List of Channels/Stimuli or measurement Markers (e.g. a-wave) to get. Can be names or indices. By default will get all.}

\item{Repeats}{Measurement repeats to get. Can be "ALL", to get all repeats, a numeric vector of length >=1 containing the repeats to keep, or a function of the form \code{function(x){x==max(x)}} indicating which repeats to keep.}

\item{Measures}{Measures to get. A character vector. Can be \code{"Voltage"}, \code{"Time"} or both (default).}

\item{Verbose}{Show warning/info messages.}
}
\value{
a data.frame.
}
\description{
Get data from an ERGExam object
}
\section{Functions}{
\itemize{
\item \code{GetAverageTraces}: Returns the average traces.
}}

\examples{
\dontrun{
# import a *.csv files exported from Diagnosys' Epsion
ERG_Experiment<-ImportEpsionMeasures("test.csv")
# get all measurements with the latest (highest number) repeat
df<-GetMeasurements(ERG_Experiment,Repeats = function(x){x==max(x)},verbose=T)
}
}
