% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggERGTrace-methods.R
\name{ggERGTrace}
\alias{ggERGTrace}
\title{Generate a ggplot2 plot for a single item/trace from an ERGExam objects}
\usage{
ggERGTrace(X, Step, Eye, Channel, Result = 1, Interactive = F)
}
\arguments{
\item{X}{A \linkS4class{ERGExam}}

\item{Step, Eye, Channel}{Vector of values for Steps, Eyes, and Channels to subset}

\item{Interactive}{Whether to return an interactive \link[plotly:ggplotly]{plotly::ggplotly}  graph}
}
\value{
A \link[ggplot2:ggplot]{ggplot2::ggplot} plot visualizing the data from a single trace from an \linkS4class{ERGExam} object
}
\description{
This method generates a \link[ggplot2:ggplot]{ggplot2::ggplot} plot for a single trace from an\linkS4class{ERGExam} objects.
}
\examples{
data(ERG)
AverageFunction(ERG, where=pairlist(Step = 3,Channel = "ERG_auto",Result = 1)) <- mean
ggERGTrace(ERG, Step = 3, Eye = "RE", Channel ="ERG_auto", Result = 1, Interactive = T) # Generate a ggplot2 plot
# the information obtained from the interactive plot can be used e.g. to update marker positions.
}
