% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_epsion.R
\name{ImportEpsion_old}
\alias{ImportEpsion_old}
\title{Import electrophysiology recordings
This function imports electrophysiology recordings from PatchMaster *.dat files ("Series") or Axon's ABF files and creates a PRecording object. Support for WinWCP files is thought in future.}
\usage{
ImportEpsion_old(filename, delim = "TAB", import = "Raw")
}
\arguments{
\item{filename}{Path to file. Currently only \href{https://www.heka.com/downloads/downloads_main.html#down_patchmaster}{Patch Master .dat} and \href{https://swharden.com/pyabf/abf2-file-format/}{Axon Binary (.abf)} files are supported.}
}
\value{
A \linkS4class{PRecording} object
}
\description{
Import electrophysiology recordings
This function imports electrophysiology recordings from PatchMaster *.dat files ("Series") or Axon's ABF files and creates a PRecording object. Support for WinWCP files is thought in future.
}
\examples{
\dontrun{
# import a PatchMaster file
tmp<-ImportPRecording("test.dat",series = 1,traces = c(1,2))
}
}
\seealso{
\linkS4class{PRecording}
}
