% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportEpsionMeasures.R
\name{ImportEpsionMeasures}
\alias{ImportEpsionMeasures}
\title{Import ERG measurements from CSV file}
\usage{
ImportEpsionMeasures(filename, sep = "\\t")
}
\arguments{
\item{filename}{Path to file. Currently supported *.csv files exported from Diagnosys' Epsion that contain at least: An area with 1) a contents table, 2) a header table, 3) a stimulus table and 4) a marker table. This function in only tested on files with the default horizontal table arrangement but should function for files with a vertical arrangement as well. Behaves otherways similar to the argument \code{file} from \link[utils:read.table]{utils::read.table()}.}

\item{sep}{The field separator character. Values on each line of the file are separated by this character. If sep = "" (the default for read.table) the separator is ‘white space’, that is one or more spaces, tabs, newlines or carriage returns.}
}
\value{
A \linkS4class{ERGExam} object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}} \cr
This function imports ERG recordings from *.csv files exported from Diagnosys' Epsion and creates an \link[=ERGExam]{ERGExam} object.
}
\examples{
\dontrun{
# import a *.csv files exported from Diagnosys' Epsion
ERG_Experiment<-ImportEpsionMeasures("test.csv")
}
}
\seealso{
\linkS4class{ERGExam}, \link[utils:read.table]{utils::read.table()}
}
