% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportEpsionProtocol.R
\name{ImportEpsionProtocol}
\alias{ImportEpsionProtocol}
\title{Import an Epsion stimulus protocol from a CSV file.}
\usage{
ImportEpsionProtocol(filename)
}
\arguments{
\item{filename}{The path to the CSV file containing the protocol data.}
}
\value{
An instance of the Protocol S4 class.
}
\description{
This function imports an Epsion stimulus protocol from a CSV file and returns it as a Protocol S4 object.
}
