% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndexOf.R
\name{IndexOf}
\alias{IndexOf}
\title{Get index of a recording}
\usage{
IndexOf(X, Step = Steps(X), Eye = Eyes(X), Channel = Channels(X), Result = 1)
}
\arguments{
\item{X}{A \linkS4class{ERGExam}}

\item{Step, Eye, Channel}{Vector of values for Steps, Eyes, and Channels to subset}
}
\value{
A numeric vector of length 1 containing the index of a recording defined by the parameters, or NULL if entry was not found.
}
\description{
This method gets the index of a recording stored in an \linkS4class{ERGExam} object.
}
\details{
This method gets the index of a recording stored in an \linkS4class{ERGExam} object.
}
