% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactiveMeasurements.R,
%   R/interactiveMeasurements_v2.R
\name{interactiveMeasurements}
\alias{interactiveMeasurements}
\title{Interactively perform measurements on an ERGExam object}
\usage{
interactiveMeasurements(X, Channel = "ERG", Eye = Eyes(X), downsample = F)

interactiveMeasurements(X, Channel = "ERG", Eye = Eyes(X), downsample = F)
}
\arguments{
\item{X}{An \linkS4class{ERGExam}}

\item{downsample}{Use peak preserving downsampling, this may accelerate potting time for large objects.}
}
\value{
An object of class 'ERGExam' with the new measurements added.

An object of class 'ERGExam' with the new measurements added.
}
\description{
This method allows interactive marker placement on \linkS4class{ERGExam} objects using  \link[shiny:shiny]{shiny:shiny-package}

This method allows interactive marker placement on \linkS4class{ERGExam} objects using  \link[shiny:shiny]{shiny:shiny-package}
}
\examples{
data(ERG)
ERG<-SetStandardFunctions(ERG)
\dontrun{
ERG<-interactiveMeasurements(ERG, Channel="OP",Eye="RE")
}
Measurements(ERG)
\dontrun{
ERG<-interactiveMeasurements(ERG, Channel="ERG")
}
Measurements(ERG)

data(ERG)
ERG<-SetStandardFunctions(ERG)
\dontrun{
ERG<-interactiveMeasurements(ERG, Channel="OP",Eye="RE")
}
Measurements(ERG)
\dontrun{
ERG<-interactiveMeasurements(ERG, Channel="ERG")
}
Measurements(ERG)

}
