% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Changelog.R
\name{LogChange}
\alias{LogChange}
\alias{LogChange,ERGExam-method}
\title{Log a Change for an ERGExam Object}
\usage{
LogChange(X)

\S4method{LogChange}{ERGExam}(X)
}
\arguments{
\item{X}{An object of class \code{ERGExam}.}
}
\value{
Invisibly returns the modified \code{ERGExam} object.
}
\description{
This internal method appends a new change entry to the \code{changelog} slot of an
\code{ERGExam} object. It checks if the function was called directly by the user
(i.e. \code{sys.nframe() == 1}). If true, it constructs a new entry with the current
date and time (formatted as "YYYYMMDD HH:MM:SS") followed by the supplied message and
appends it to the changelog.
}
\section{Methods (by class)}{
\itemize{
\item \code{LogChange(ERGExam)}: Append a new log entry to the changelog slot if called directly by the user.

}}
\keyword{internal}
