% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeExample.R
\name{makeExampleERGSteps}
\alias{makeExampleERGSteps}
\title{Create Example ERGSteps Object}
\usage{
makeExampleERGSteps()
}
\value{
An instance of the ERGSteps class containing example data and metadata.
}
\description{
This function creates an example ERGSteps object with mock data and metadata.
}
\examples{
example_erg_steps <- makeExampleERGSteps()

}
