% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MINI.R
\name{MINI}
\alias{MINI}
\title{Report MINI Information for an ERGExam object}
\usage{
MINI(X, ValidateTaxa = TRUE, tidy = FALSE)
}
\arguments{
\item{X}{An \linkS4class{ERGExam}, \linkS4class{ERGMeasurements}, or \linkS4class{ERGElectrode} object.}

\item{ValidateTaxa}{Should taxonomy data be validated? Default: \code{TRUE}.}

\item{tidy}{Logical. Should the output be returned in a tidy vertical format? Default: \code{FALSE}.}
}
\value{
A \code{data.frame} with one row, comprising the extracted MINI exam and subject information,
including a computed \code{Age} field.
}
\description{
This method extracts relevant MINI information as proposed by Gibson et al (https://doi.org/10.1038/npre.2009.1720.2) from the \code{ExamInfo} and \code{SubjectInfo} slots of an \code{ERGExam} object.
It returns a one-row \code{data.frame} containing exam and subject details.
Additionally, if both \code{ExamDate} (from \code{ExamInfo}) and \code{DOB} (from \code{SubjectInfo}) are available,
the subject's \code{Age} is computed as the difference between these dates in years.
}
\examples{
data(ERG)
MINI(ERG)
}
