% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0ERGElectrode-class.R
\name{newERGElectrode}
\alias{newERGElectrode}
\title{Create an instance of the ERGElectrode class}
\usage{
newERGElectrode(Name, Type, Position, Impedance)
}
\arguments{
\item{Name}{Character. The identifier of the electrode (e.g. "Ground").}

\item{Type}{Character. The type of the electrode (e.g. "Platinum").}

\item{Position}{Character. The position of the electrode ("E.g. Tail").}

\item{Impedance}{Numeric of class units. The impedance of the electrode. Unit must be convertible to Ohm.}
}
\value{
An object of class \code{ERGElectrode}.
}
\description{
This function creates an instance of the \linkS4class{ERGElectrode} class.
}
\seealso{
\linkS4class{ERGElectrode}
}
