% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{PlotExam}
\alias{PlotExam}
\title{Plot an ERGExam object.}
\usage{
PlotExam(X, labelby = "Intensity", facetby = "Eye+Channel")
}
\arguments{
\item{X}{An ERGExam object.}

\item{labelby}{A character string specifying the variable to color by (default: "Intensity").}

\item{facetby}{TBC}
}
\value{
A ggplot2 plot of the ERGExam data.
}
\description{
This function creates a plot of an ERGExam object, including line plots of
stimulus response curves for different label categories.
}
\examples{
# Example usage:
exam_plot <- PlotExam(erg_exam)
print(exam_plot)

}
