% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotList.R
\name{PlotRecordings}
\alias{PlotRecordings}
\title{Plot ERG recordings}
\usage{
PlotRecordings(
  List,
  Background = "DA",
  Type = "Flash",
  Channel = "ERG",
  wrap_by = "Channel",
  scales = "free_y"
)
}
\arguments{
\item{List}{A list of ERG exams.}

\item{Background}{Background condition for the exams.}

\item{Type}{Type of exam (e.g., "Flash").}

\item{Channel}{The channel to plot (e.g., "ERG").}

\item{wrap_by}{Wrapping parameter for facetting ("Channel" or NULL).}
}
\value{
A ggplot2 plot object.
}
\description{
This function generates a plot of ERG recordings.
}
\examples{
# Example usage:
data <- list(ERGExam1, ERGExam2)
PlotRecordings(data, Background = "DA", Type = "Flash", Channel = "ERG")

}
