% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotList.R
\name{PlotStepSequence}
\alias{PlotStepSequence}
\title{Plot step sequence for ERG exams}
\usage{
PlotStepSequence(
  List,
  Background = "DA",
  Type = "Flash",
  Channel = NULL,
  Markers = c("N1", "P1"),
  wrap_by = "Channel"
)
}
\arguments{
\item{List}{A list of ERG exams.}

\item{Background}{Background condition for the exams.}

\item{Type}{Type of exam (e.g., "Flash").}

\item{Channel}{The channel to plot (e.g., "ERG").}

\item{Markers}{Vector of markers to include in the plot (e.g. c("a","B)).}

\item{wrap_by}{Wrapping parameter for facetting ("Channel" or NULL).}
}
\value{
A ggplot2 plot object.
}
\description{
This function generates a plot of step sequence data for ERG exams.
}
\examples{
# Example usage:
data <- list(ERGExam1, ERGExam2)
PlotStepSequence(data, Background = "DA", Type = "Flash", Channel = "ERG")

}
