% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetExamSubject-methods.R
\name{Set}
\alias{Set}
\alias{Amplifier<-}
\alias{Disease<-}
\alias{Development<-}
\alias{Electrodes<-}
\alias{Genotype<-}
\alias{Genus<-}
\alias{ID<-}
\alias{Investigator<-}
\alias{Label<-}
\alias{ProtocolName<-}
\alias{ProtocolVersion<-}
\alias{RecMode<-}
\alias{Recorder<-}
\alias{Species<-}
\alias{Stimulator<-}
\alias{Strain<-}
\title{Set Methods for Exam and Subject information for the \linkS4class{ERGExam} object}
\usage{
Amplifier(X) <- value

Disease(X) <- value

Development(X) <- value

Electrodes(X) <- value

Genotype(X) <- value

Genus(X) <- value

ID(X) <- value

Investigator(X) <- value

Label(X) <- value

ProtocolName(X) <- value

ProtocolVersion(X) <- value

RecMode(X) <- value

Recorder(X) <- value

Species(X) <- value

Stimulator(X) <- value

Strain(X) <- value
}
\arguments{
\item{X}{An \linkS4class{ERGExam}, \linkS4class{ERGMeasurements}, or \linkS4class{ERGElectrode} object.}

\item{value}{The new value to be set for the specified field.}
}
\value{
An updated version of the \linkS4class{ERGExam} object.
}
\description{
These methods are used to set exam and subject information from \linkS4class{ERGExam} or \linkS4class{ERGMeasurements} objects.
}
\section{Functions}{
\itemize{
\item \code{Amplifier(X) <- value}: Sets the Amplifier field.

\item \code{Disease(X) <- value}: Sets the Disease information.

\item \code{Development(X) <- value}: Sets the Development information.

\item \code{Electrodes(X) <- value}: Sets the Electrodes field.

\item \code{Genotype(X) <- value}: Sets the Genotype information.

\item \code{Genus(X) <- value}: Sets the Genus information.
Uses \code{validateTaxa} to ensure that the provided value is a recognized genus.

\item \code{ID(X) <- value}: Sets the ID information.

\item \code{Investigator(X) <- value}: Sets the Investigator.

\item \code{Label(X) <- value}: Sets the Label information.

\item \code{ProtocolName(X) <- value}: Sets the ProtocolName.

\item \code{ProtocolVersion(X) <- value}: Sets the ProtocolVersion.

\item \code{RecMode(X) <- value}: Sets the Recording Mode.

\item \code{Recorder(X) <- value}: Sets the Recorder field.

\item \code{Species(X) <- value}: Sets the Species information.
Uses \code{validateTaxa} to ensure that the provided value is a recognized species.

\item \code{Stimulator(X) <- value}: Sets the Stimulator information.

\item \code{Strain(X) <- value}: Sets the Strain information.

}}
\examples{
data(ERG)  # assuming data(ERG) gives a valid ERGExam object

# Set Amplifier
Amplifier(ERG) <- "Amp_001"
Amplifier(ERG)

# Set Disease
Disease(ERG) <- "None"
Disease(ERG)

# Set Development
Development(ERG) <- "Adult"
Development(ERG)

# Set Electrodes
# (Assume electrodes_list is a named list of ERGElectrode objects where names match each element's Name(ERGElectrode))
# Electrodes(ERG) <- electrodes_list

# Set Genotype
Genotype(ERG) <- "Wild-type"
Genotype(ERG)

# Set Genus (using a recognized genus, e.g., "Homo")
Genus(ERG) <- "Mus"
Genus(ERG)

# Set ID
ID(ERG) <- "Subject_001"
ID(ERG)

# Set Investigator
Investigator(ERG) <- "Dr. Smith"
Investigator(ERG)

# Set Label (if applicable)
Label(ERG) <- "Fluorescent"
Label(ERG)

# Set ProtocolName
ProtocolName(ERG) <- "ERG_Protocol_1"
ProtocolName(ERG)

# Set ProtocolVersion
ProtocolVersion(ERG) <- "1.0"
ProtocolVersion(ERG)

# Set RecMode
RecMode(ERG) <- "ModeA"
RecMode(ERG)

# Set Recorder
Recorder(ERG) <- "Rec_001"
Recorder(ERG)

# Set Species (using a recognized species, e.g., "sapiens")
Species(ERG) <- "musculus"
Species(ERG)

# Set Stimulator
Stimulator(ERG) <- "Stim_001"
Stimulator(ERG)

# Set Strain
Strain(ERG) <- "C57BL/6"
Strain(ERG)
}
