% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0ERGProtocol.R
\name{show.Protocol}
\alias{show.Protocol}
\title{Show method for Protocol class}
\usage{
\S4method{show}{Protocol}(object)
}
\arguments{
\item{object}{An instance of the Protocol class.}

\item{...}{Additional arguments (not used).}
}
\description{
Show method for Protocol class
}
