% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{show}
\alias{show}
\title{Show}
\arguments{
\item{object}{An S4 object of type \link[=ERGExam]{ERGExam}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}} \cr
Default method \code{show} for \link[=ERGExam]{ERGExam}.
}
