% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Subset.R
\name{Subset}
\alias{Subset}
\title{Subset from ERGExam Object}
\usage{
\S4method{Subset}{ERGExam}(
  X,
  Time = NULL,
  TimeExclusive = FALSE,
  Repeats = NULL,
  Raw = TRUE,
  Step = Steps(X),
  Eye = Eyes(X),
  Channel = Channels(X),
  ...
)
}
\arguments{
\item{Step, Eye, Channel}{Vector of values for Steps, Eyes, and Channels to subset}
}
\description{
This method subsets an \code{ERGExam} object into a new object of the same class.
}
\details{
The \code{Subset} function creates a new \code{ERGExam}  object containing a subset of the data from the original object, based on the provided parameters.
}
\seealso{
\link[EPhysData:Subset]{EPhysData::Subset}
}
