% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateMeasurements.R
\name{UpdateMeasurements-method}
\alias{UpdateMeasurements-method}
\alias{UpdateMeasurements,ERGExam-method}
\title{UpdateMeasurements Method for ERGExam Class}
\usage{
\S4method{UpdateMeasurements}{ERGExam}(X, values)
}
\arguments{
\item{X}{An \code{EPhysData} object}

\item{values}{A data.frame containing new measurement values with columns 'Recording', 'Name', 'Voltage', and 'Time'.}
}
\value{
An updated 'ERGExam' object with the measurements replaced or added based on the 'values' data.frame.
}
\description{
This method updates the measurements slot in an 'ERGExam' object with new data.
}
\details{
This method is used to update the measurements of an 'ERGExam' object with new data. The method first checks that the 'values' data.frame contains the required columns. It then merges the existing measurements with the new data based on the 'Recording' and 'Name' columns. If a matching entry is found, the 'Voltage' and 'Time' values in the 'ERGExam' object are updated with the corresponding values from 'values'. If no matching entry is found, a new entry is added to the measurements.
}
