% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateMeasurements.R
\name{UpdateMeasurements}
\alias{UpdateMeasurements}
\alias{[<-,ERGExam-method}
\title{UpdateMeasurements Method for ERGExam Class}
\usage{
UpdateMeasurements(X, values)

\S4method{[}{ERGExam}(x, i, j, ...) <- value
}
\arguments{
\item{X}{A \linkS4class{ERGExam}}

\item{values}{FIXME A data.frame containing new measurement values with columns 'Recording', 'Name', 'Voltage', and 'Time'.}

\item{...}{Additional parameters.
These additional parameters provide flexibility for users to specify the Recording the measurement should be assigned to by either giving the index, or parameters defining the recording.
- \code{Marker.Name}: A character string specifying the marker name. Mandatory for setting the marker value.
- \code{Recording}: The index of the recording where the marker value will be set. If not provided, the parameters \code{Step}, \code{Eye}, \code{Channel}, and \code{Result} must be provided instead.
- \code{Step}: The step index. Required if \code{Recording} is not provided.
- \code{Eye}: The eye identifier. Required if \code{Recording} is not provided.
- \code{Channel}: The channel identifier. Required if \code{Recording} is not provided.
- \code{Result}: The result index. Required if \code{Recording} is not provided.}

\item{value}{The value to set. This can be either a numeric value giving the time position of the measurement or a pair list containing 'Voltage', 'Time', and optionally 'Relative' entries.}

\item{Time}{The time on trace at which the measurement was taken}
}
\value{
An updated \link[=ERGExam]{ERGExam} object with the measurements replaced or added based on the data.frame provided by the \code{values} attribute.
}
\description{
This method updates the measurements slot in an \link[=ERGExam]{ERGExam} object with new data.
}
\details{
This method is used to update the measurements of an \link[=ERGExam]{ERGExam} object with new data. The method first checks that the 'values' data.frame contains the required columns. It then merges the existing measurements with the new data based on the 'Recording' and 'Name' columns. If a matching entry is found, the 'Voltage' and 'Time' values in the 'ERGExam' object are updated with the corresponding values from 'values'. If no matching entry is found, a new entry is added to the measurements.
}
\section{Functions}{
\itemize{
\item \code{UpdateMeasurements()}: Flexibly updates the Measurements slot of an \link[=ERGExam]{ERGExam} object. Replaces marker times and values where already existent and adds them where not.

\item \code{`[`(ERGExam) <- value}: Update or add a line to the Measurements slot in an ERGExam object: Either a new time for a given marker can be provided and the corresponding value is extracted automatically from the trace or a combination of time, voltage and information on whether the measurement was performed relative to another marker can be provided.

}}
