% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateProcessingMethods.R
\name{UpdateProcessingMethods}
\alias{UpdateProcessingMethods}
\alias{FilterFunction<-,ERGExam-method}
\alias{Rejected<-,ERGExam-method}
\alias{AverageFunction<-,ERGExam-method}
\alias{SetStandardFunctions}
\title{Set processing functions for ERGExam objects}
\usage{
\S4method{FilterFunction}{ERGExam}(X, where) <- value

\S4method{Rejected}{ERGExam}(X, where) <- value

\S4method{AverageFunction}{ERGExam}(X, where) <- value

SetStandardFunctions(
  X,
  Stimulus.type.names = pairlist(Flash = "Flash", Flicker = "Flicker")
)
}
\arguments{
\item{X}{An \linkS4class{ERGExam} object}

\item{where}{A \link[base:pairlist]{base::pairlist} allowing select criteria for which recordings the functions should be set. Tags/Keys in the pairlist must represent valid column names of \link{Metadata} or\link{StimulusTable}}

\item{value}{A value (usually a function) to set.}

\item{Stimulus.type.names}{A \link[base:pairlist]{base::pairlist} specifying the names identifying the different stimulus types, e.g., \code{Flash="Flash"} or \code{Flash="Blitz"}.}
}
\value{
An updated ERGExan object.

#' @examples
data(ERG)
AverageFunction(ERG,where=pairlist(Step=1))<-median
stop(GetData(updated_ERG,raw=F))

updated_ERG
}
\description{
These methods are used to set filter, rejection and averaging functions for \link{ERGExam} objects.
They allow setting the functions to a subset of recordings based on the specified conditions.
}
\section{Functions}{
\itemize{
\item \code{FilterFunction(ERGExam) <- value}: Update the FilterFunction for all Recordings in an \linkS4class{ERGExam}, or only those slected using \code{where}.

\item \code{Rejected(ERGExam) <- value}: Update the Rejected for all Recordings in an \linkS4class{ERGExam}, or only those slected using \code{where}.

\item \code{AverageFunction(ERGExam) <- value}: Update the AverageFunction for all Recordings in an \linkS4class{ERGExam}, or only those slected using \code{where}.

\item \code{SetStandardFunctions()}: This method is used to set standard functions for processing \linkS4class{ERGExam} data. It defines default functions for averaging, filtering, and signal rejection based on the stimulus type.

}}
\seealso{
\link[EPhysData:Get_Set_EPhysData]{EPhysData::Get_Set_EPhysData}, \link[EPhysMethods:autoreject.by.distance]{EPhysMethods::autoreject.by.distance}, \link[EPhysMethods:autoreject.by.signalfree]{EPhysMethods::autoreject.by.signalfree}, \link[EPhysMethods:filter.bandpass]{EPhysMethods::filter.bandpass}, \link[EPhysMethods:filter.detrend]{EPhysMethods::filter.detrend},
}
