% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MINI.R
\name{validateTaxa}
\alias{validateTaxa}
\title{Validate Genus/Species combination using taxize (internal)}
\usage{
validateTaxa(variable, rank_query)
}
\arguments{
\item{genus}{A character string specifying the genus.}

\item{species}{A character string specifying the species.}
}
\value{
\code{TRUE} if the taxon is recognized, \code{FALSE} if not, and \code{NA} if \pkg{taxize} is not installed.
}
\description{
This function checks whether the provided Genus and Species correspond to a recognized taxon.
If the \pkg{taxize} package is available, it uses \code{taxize::get_uid} to perform the lookup.
Otherwise, it issues a CLI warning and skips the taxonomic validation.
}
\keyword{internal}
