% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClearMeasurements-method.R
\name{ClearMeasurements}
\alias{ClearMeasurements}
\title{Clear Computed Slots from an ERGExam object}
\usage{
ClearMeasurements(X)
}
\arguments{
\item{X}{A \linkS4class{ERGExam}}
}
\value{
An ERGExam object with the Results slot cleared.
}
\description{
Clears the Measurements slots in an  \link[=ERGExam]{ERGExam} object, respectively, by removing any existing data.
}
