% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERGtools2-package.R
\docType{package}
\name{.Overview}
\alias{ERGtools2}
\alias{ERGtools2-package}
\title{Structured Index and Introduction for the ERGtools2 R Package}
\description{
This package contains an environment for working with electroretinogram data. It contains an import method for Diagnosys Espion data, but allows reading in of data also from other manufacturers with limited coding effort. Standard procedures like averaging, subsetting an visualization of individual exams are supported.
This package provides the  \linkS4class{ERGExam} class that stores data from a single ERG examination. These may include different recording channels (ERG, OP, VEP, ...), or sequential recordings in repsonse to different stimulus paradigms.
It is usually generated from imported raw data using \link[=newERGExam]{newERGExam}. Data acuired from Diagnosys Espion can be imported directly using \link[=ImportEspion]{ImportEspion}.\cr\cr
}
\details{
The class \linkS4class{ERGExam} expands \link[EPhysData:EPhysSet]{EPhysData::EPhysSet}, so \strong{Methods that work on \link[EPhysData:EPhysSet]{EPhysData::EPhysSet} can be applied to \linkS4class{ERGExam} as well. See: \link[EPhysData:EPhysData-package]{EPhysData::EPhysData-package}}.
}
\section{Standard workflow}{


After setting the functions, the accession methods inherited from \link[EPhysData:EPhysSet-class]{EPhysData::EPhysSet-class} like \link[EPhysData:as.data.frame-method]{EPhysData::as.data.frame-method} and \link[EPhysData:GetData]{EPhysData::GetData} with the argument \code{Raw=F} can be used to return the processed data.\cr\cr
}

\section{Object creation}{

\itemize{
\item \link[=newERGExam]{newERGExam} and \link[=ImportEspion]{ImportEspion} for \link[=ERGExam]{ERGExam} objects. \cr
\item \link[methods:new]{methods:new} and \link[=ImportEspionProtocol]{ImportEspionProtocol} for \link[=ERGProtocol]{ERGProtocol} objects. Experimental. \cr
\item \link[methods:new]{methods:new} for \link[=ERGMeasurements-class]{ERGMeasurements-class} objects. \cr
\item data(ERG) (\link[=ERGExam-data]{.SampleERGExam}) Load example ERG recording \cr\cr
\item data(Measurements.data) (\link[=ERGMeasurements-data]{.SampleERGMeasurements}) Load example Measurements data \cr\cr
}
}

\section{Accession methods}{

\itemize{
\item \link[=Where]{Where} Returns the recording index for those recordings matching the given criteria. \cr\cr
\item \link[=Subset]{Subset} This method subsets an (for \link[=ERGExam]{ERGExam} object into a new object of the same class.
\item \link[=as.data.frame]{as.data.frame} Returns data frame representing the \link[=ERGExam]{ERGExam} or \link{ERGProtocol} object in long format. When used with the argument \code{Raw = F} on an\link[=ERGExam]{ERGExam} process (i.e. filtered, averaged) data is returned. See also: \link[EPhysData:as.data.frame-method]{EPhysData::as.data.frame-method}. \cr\cr
\item \link[=DOB]{DOB}, \link[=ExamDate]{ExamDate}, \link[=GroupName]{GroupName}, \link[=ProtocolName]{ProtocolName}: Returning information on the exam and the examined subject.
\item \link[=Eyes]{Eyes}, \link[=Steps]{Steps}, \link[=Channels]{Channels}, \link[=Results]{Results}: Returning information on the Recordings contained in the dataset.
\item \link[=StimulusTableMethods]{StimulusTableMethods} Stimulus(): Returns selected rows of a stimulus table.
\item \link[=StimulusTableMethods]{StimulusTableMethods} StimulusDescription(), StimulusIntensity(), StimulusBackground(), StimulusType(). \cr\cr
\item \link[=MarkerNames]{MarkerNames}
\item \link[=Markers]{Markers}
\item \link[=Measurements]{Measurements} Returns the Measurements table.
}
}

\section{Processing}{

\itemize{
\item \link[=FilterFunction<-]{FilterFunction<-} Update the FilterFunction for all Recordings in an \linkS4class{ERGExam}, or only those slected using \code{where}.
\item \link[=Rejected<-]{Rejected<-} Update the Rejected for all Recordings in an \linkS4class{ERGExam}, or only those slected using \code{where}.
\item \link[=AverageFunction<-]{AverageFunction<-} Update the AverageFunction for all Recordings in an \linkS4class{ERGExam}, or only those slected using \code{where}.
\item \link[=SetStandardFunctions]{SetStandardFunctions} This method is used to set standard functions for processing \linkS4class{ERGExam} data. It defines default functions for averaging, filtering, and signal rejection based on the stimulus type.
\item \link[=AutoPlaceMarkers]{AutoPlaceVEP} Automatically sets markers depending on the channel (E.g. ERG, VEP, OP,...) and stimulus type (Flash, FLicker).
\item \link[=AutoPlaceAB]{AutoPlaceAB} Place the a and B waves on Flash ERG data stored in an an \link[EPhysData:EPhysData-class]{EPhysData::EPhysData-class} object.
\item \link[=AutoPlaceFlicker]{AutoPlaceFlicker} Place the N1 and P1 markers and determines 1/frequency (period) for Flicker ERG data stored in an an \link[EPhysData:EPhysData-class]{EPhysData::EPhysData-class} object.
\item \link[=AutoPlaceVEP]{AutoPlaceVEP} Place the P1, N1 and P2 markers for Flash VEP data stored in an an \link[EPhysData:EPhysData-class]{EPhysData::EPhysData-class} object.
\item \link[=CheckAvgFxSet]{CheckAvgFxSet}
\item \link[=interactiveMeasurements]{interactiveMeasurements}: Interactive visual placement of markers.
}
}

\section{Merging and other object manipulation}{

\itemize{
\item \link[=MergeERGExams]{MergeERGExams} (for \link[=ERGExam]{ERGExam} objects) \cr\cr
\item \link[=Measurements<-]{Measurements<-} Add, update or remove Measurements from an \link[=ERGExam]{ERGExam} or \link[=ERGMeasurements]{ERGMeasurements-class} object.
\item \link[=DropMarker]{DropMarker}
\item \link[=AddMarker]{DropMarker} \cr\cr
\item \link[=StimulusTableMethods]{StimulusTableMethods} StimulusDescription()<-, StimulusIntensity()<-, StimulusBackground()<-, StimulusType()<-\cr\cr
\item \link[=UpdateChannelNames]{UpdateChannelNames} Update or replace channel names.
\item \link[=ClearMeasurements]{ClearMeasurements}  Clear the Measurements slots in an \link[=ERGExam]{ERGExam} object.
}
}

\section{Plot methods}{

\itemize{
\item \link[=ggERGTrace]{ggERGTrace} Generate a \link[ggplot2:ggplot]{ggplot2::ggplot} plot for a single trace from an \linkS4class{ERGExam} objects.
\item \link[=ggERGExam]{ggERGExam}	Plot a complete \linkS4class{ERGExam} object.
\item \link[=ggIntensitySequence]{ggIntensitySequence}	Uses \link[ggplot2:ggplot]{ggplot2::ggplot} to plot intensity sequence for ERG exams
\item \link[=ggStepSequence]{ggStepSequence}	Uses \link[ggplot2:ggplot]{ggplot2::ggplot}  to plot step sequence (i.e. sequential recordings within a single protocol) for ERG exams
\item \link[=ggPlotRecordings]{ggPlotRecordings}	Uses ggplot2 to plot ERG traces from multiple ERGExam objects
\item \link[=interactiveMeasurements]{interactiveMeasurements}: Interactive visual placement of markers.#'
}
}

\examples{
# a typical workflow
data(ERG) # load example data
StimulusTable(ERG) # have a look whats inside
Metadata(ERG)
ERG<-SetStandardFunctions(ERG)
ggERGTrace(ERG, where = list( Step = as.integer(3), Eye = "RE", Channel ="ERG_auto", Result = as.integer(1))) # pick one and have a look at the traces as imported

ERG <- AutoPlaceMarkers(ERG, Channel.names = pairlist(ERG = "ERG_auto")) # automatically place markers

ggERGTrace(ERG, where = list( Step = as.integer(3), Eye = "RE", Channel ="ERG_auto", Result = as.integer(1))) # pick one and have a look at the traces as imported

}
\author{
\href{https://www.lindnerlab.de}{Moritz Lindner}
}
