% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0ERGExam-class.R
\docType{class}
\name{ERGExam}
\alias{ERGExam}
\title{ERGExam Class}
\description{
A class representing an ERG (Electroretinogram) exam with associated data and attributes. This class extends the \link[EPhysData:EphysSet]{EPhysData::EphysSet} object and all methods valid for \link[EPhysData:EphysSet]{EPhysData::EphysSet} can also be applied to \code{ERGExam} objects.
}
\section{Slots}{

\describe{
\item{\code{Data}}{Data A list of \link{EPhysData} objects. Each item containing a recording in respsonse to a particular Stimulus ("Step"), from a particular eye and data Channel (e.g. ERG or OP)-}

\item{\code{Metadata}}{A data frame containing metadata information associated with the data, each row corresponds to one list item.
\describe{
\item{Step}{A character vector containing the steps associated with the data.}
\item{Eye}{A character vector containing the possible values "RE" (right eye) and "LE" (left eye).}
\item{Channel}{A character vector containing the unique names of the third level of \code{Data}.}
}}

\item{\code{Stimulus}}{A data frame containing stimulus information associated with the data.
\describe{
\item{Step}{A character vector containing the steps associated with the stimulus.}
\item{Description}{A character vector describing the stimuli.}
\item{Intensity}{An integer vector representing the intensity of the stimulus (unitless).}
\item{Background}{A character vector describing the adaptation state of the stimulus (DA or LA).}
\item{Type}{A character vector describing the type of the stimulus (e.g. Flash or Flicker).}
}}

\item{\code{Averaged}}{TRUE if the object contains averaged data, FALES indicates object contains raw traces.}

\item{\code{Measurements}}{A data frame containing measurements information associated with the data.
\describe{
\item{Recording}{A numeric vector representing recording identifiers corresponding to the data.}
\item{Name}{A character vector containing the names of the measurements.}
\item{Time}{A numeric vector representing time measurements (seconds).}
\item{Value}{A numeric vector representing measurement values (unitless).}
}}

\item{\code{ExamInfo}}{A list containing exam-related information.
\describe{
\item{ProtocolName}{A character vector indicating the name of the protocol.}
\item{Version}{Optional: A character vector indicating the version of the protocol.}
\item{ExamDate}{A \code{POSIXct} times tamp representing the date of the exam.}
\item{Filename}{Optional: A character vector indicating the filename associated with the exam data.}
\item{RecMode}{Optional: A character vector indicating the recording mode during the exam.}
\item{Investigator}{Optional: A character vector indicating the name of the investigator conducting the exam.}
}}

\item{\code{SubjectInfo}}{A list containing subject-related information.
\describe{
\item{Subject}{A character vector indicating the name of the patient.}
\item{DOB}{A \code{Date} object indicating the date of birth of the patient.}
\item{Gender}{Optional: A character vector indicating the gender of the patient.}
\item{Group}{Optional: A character vector indicating the group to which the patient belongs.}
}}

\item{\code{Imported}}{A \code{POSIXct} timestamp indicating when the object was imported.}
}}

\seealso{
\link[EPhysData:EPhysData-package]{EPhysData::EPhysData-package}\link[EPhysData:EPhysData-class]{EPhysData::EPhysData-class} \link[EPhysData:EphysSet-class]{EPhysData::EPhysSet-class}
}
