% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetData.R
\name{GetMeasurements}
\alias{GetMeasurements}
\title{Get measurements data from an ERGExam object}
\usage{
GetMeasurements(
  X,
  Channels = GetChannelNames(X),
  Stimuli = GetStimulusNames(X),
  Repeats = "ALL",
  Markers = GetMarkerNames(X),
  Measures = c("Voltage", "Time"),
  Verbose = T
)
}
\value{
An \linkS4class{ERGExam} object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}} \cr
This function subsets \linkS4class{ERGExam} objects by \var{Channels}, \var{Stimuli}, \var{Repeats}, measurement \var{Markers} or \var{Measures}.
}
\examples{
\dontrun{
# import a *.csv files exported from Diagnosys' Epsion
ERG_Experiment<-ImportEpsionMeasures("test.csv")
# get all measurements with the latest (highest number) repeat
df<-GetMeasurements(ERG_Experiment,Repeats = function(x){x==max(x)},verbose=T)
}
}
