% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggPlotList.R
\name{ggPlotRecordings}
\alias{ggPlotRecordings}
\title{Uses ggplot2 to plot ERG traces from multiple ERGExam objects}
\usage{
ggPlotRecordings(
  List,
  Background = "DA",
  Type = "Flash",
  Channel = "ERG",
  wrap_by = "Channel",
  scales = "free_y"
)
}
\arguments{
\item{List}{A list of ERG exams.}

\item{Background}{Background condition for the exams.}

\item{Type}{Type of exam (e.g., "Flash").}

\item{Channel}{The channel to plot (e.g., "ERG").}

\item{wrap_by}{Wrapping parameter for facetting ("Channel" or NULL).}
}
\value{
A ggplot2 plot object.
}
\description{
This function generates a \link[ggplot2:ggplot]{ggplot2:ggplot} plot of ERG traces from multiple ERGExam objects.
}
\examples{
# Example usage:
data(ERG)
ERG<-SetStandardFunctions(ERG)
ERG <- AutoPlaceMarkers(ERG)
data <- list(ERG, ERG)
ggPlotRecordings(data, Background = "DA", Type = "Flash", Channel = "ERG")

}
