% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportEpsion.R
\name{ImportEpsion}
\alias{ImportEpsion}
\title{Import ERG data and measurements from an Eosion CSV file}
\usage{
ImportEpsion(
  filename,
  sep = "\\t",
  Import = list("Averaged", "Measurements"),
  Protocol = NULL
)
}
\arguments{
\item{filename}{Path to the file. Currently supported \code{'.csv'} or \code{'.txt'} files exported from Diagnosys' Epsion that contain at least:
An area with 1) a contents table, 2) a header table, 3) a stimulus table, and 4) a marker table.
This function is only tested on files with the default horizontal table arrangement but should function for files with a vertical arrangement as well.
Behaves similarly to the argument \code{file} from \link[utils:read.table]{utils::read.table()}.}

\item{sep}{The field separator character. Values on each line of the file are separated by this character.
If sep = "" (the default for read.table), the separator is ‘white space’, which includes spaces, tabs, newlines, or carriage returns.}

\item{Import}{A list of character vectors specifying which parts of the data to import.
Possible elements are "Raw" - for importing the raw recordings, "Averaged" - for importing the averaged recordings ("Results"), and "Measurements" - for importing the measured markers. Either of "Raw" or "Averaged" must be selected.}

\item{Protocol}{An S4 object of class \link[Protocol]{Protocol()} or a list thereof.}
}
\value{
A \linkS4class{ERGExam} object.
}
\description{
This function imports ERG recordings from *.csv files exported from Diagnosys' Epsion and creates an \link[=ERGExam]{ERGExam} object.
}
\examples{
\dontrun{
# Import a *.csv file exported from Diagnosys' Epsion
ERG_Experiment <- ImportEpsion("test.csv")
}

}
\seealso{
\linkS4class{ERGExam} \link[Protocol]{Protocol()}
}
