% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MergeERGExams.R
\name{MergeERGExams}
\alias{MergeERGExams}
\title{Merge ERGExams}
\usage{
MergeERGExams(X, examX)
}
\arguments{
\item{X}{A \linkS4class{ERGExam}}

\item{examX}{An ERGExam object or a list of ERGExam objects to be merged with \code{X}.}
}
\value{
An ERGExam object representing the merged data. the file names, protocol names and dates from the individual files merged are stored as additional columns in the Stimulus table. Use \linkS4class{StimulusTable}(X, full=T) to see those.
}
\description{
Merges two or more ERGExam objects into a single ERGExam object.
}
\examples{
# Merge two ERGExams
merged_exam <- mergeERGExams(exam1, exam2)

# Merge a list of ERGExams
exam_list <- list(exam2, exam3, exam4)
merged_exam_list <- mergeERGExams(exam1, exam_list)

}
