% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0ERGExam-class.R
\name{newERGExam}
\alias{newERGExam}
\title{Create an instance of the ERGExam class}
\usage{
newERGExam(
  Data,
  Metadata,
  Stimulus,
  Averaged = FALSE,
  Measurements = data.frame(Recording = numeric(), Name = character(), Time =
    as_units(integer(), "s"), Value = as_units(integer(), unitless), stringsAsFactors =
    FALSE),
  ExamInfo,
  SubjectInfo
)
}
\arguments{
\item{Data}{A list of \linkS4class{EPhysData::EPhysData} objects.}

\item{Metadata}{A data frame containing metadata information associated with the data, each row corresponds to one list item.
\describe{
\item{Step}{A character vector containing the steps associated with the data.}
\item{Eye}{A character vector containing the possible values "RE" (right eye) and "LE" (left eye).}
\item{Channel}{A character vector containing the unique names of the third level of \code{Data}.}
}}

\item{Stimulus}{A data frame containing stimulus information associated with the data.
\describe{
\item{Step}{A character vector specifying the steps associated with the stimulus.}
\item{Name}{A character vector specifying the names of the stimuli.}
\item{Description}{A character vector describing the stimuli.}
\item{Adaptation}{A character vector describing the adaptation state of the stimulus.}
\item{Background}{An integer vector representing the background intensity of the stimulus (unitless).}
\item{Intensity}{An integer vector representing the intensity of the stimulus (unitless).}
}}

\item{Averaged}{A list of averaged data.}

\item{Measurements}{A data frame containing measurements information associated with the data.
\describe{
\item{Recording}{A numeric vector specifying recording identifiers corresponding to the data.}
\item{Name}{A character vector specifying the names of the measurements.}
\item{Time}{A numeric vector specifying time measurements (seconds).}
\item{Value}{A numeric vector specifying measurement values (unitless).}
}}

\item{ExamInfo}{A list containing exam-related information.
\describe{
\item{ProtocolName}{A character vector indicating the name of the protocol.}
\item{Version}{Optional: A character vector indicating the version of the protocol.}
\item{ExamDate}{A \code{POSIXct} timestamp representing the date of the exam.}
\item{Filename}{Optional: A character vector indicating the filename associated with the exam data.}
\item{RecMode}{Optional: A character vector indicating the recording mode during the exam.}
\item{Investigator}{Optional: A character vector indicating the name of the investigator conducting the exam.}
}}

\item{SubjectInfo}{A list containing subject-related information.
\describe{
\item{Subject}{A character vector indicating the name of the patient.}
\item{DOB}{A \code{Date} object indicating the date of birth of the patient.}
\item{Gender}{Optional: A character vector indicating the gender of the patient.}
\item{Group}{Optional: A character vector indicating the group to which the patient belongs.}
}}
}
\value{
An object of class \code{ERGExam}.
}
\description{
This function creates an instance of the \code{ERGExam} class
with the specified data and attributes.
}
\examples{
# Create example data and metadata
Data <-
  list(
    makeExampleEPhysData(nrows = 10, replicate_count = 3),
    makeExampleEPhysData(nrows = 10, replicate_count = 3),
    makeExampleEPhysData(nrows = 5, replicate_count = 6),
    makeExampleEPhysData(nrows = 5, replicate_count = 6)
  )  # List of EPhysData objects
Metadata <-
  data.frame(
    Step = c("A1", "A1", "A2", "A2"),
    Eye = c("RE", "LE", "LE", "LE"),
    Channel = c("Ch1", "Ch1", "Ch1", "Ch2")
  )
Stimulus <-
  data.frame(Step = c("A1", "A2"),
             Name = c("Stim1", "Stim2"))  # Example stimulus data
ExamInfo <-
  list(
    ProtocolName = "ERG Protocol",
    Version = "1.0",
    ExamDate = as.POSIXct("2023-08-14"),
    Filename = "exam_data.csv"
  )
SubjectInfo <-
  list(
    Subject = "John Doe",
    DOB = as.Date("1990-05-15"),
    Gender = "Male",
    Group = "Control"
  )
ergExam <-
  newERGExam(
    Data = Data,
    Metadata = Metadata,
    Stimulus = Stimulus,
    ExamInfo = ExamInfo,
    SubjectInfo = SubjectInfo
  )

}
\seealso{
\code{\link{ERGExam-class}}
}
