% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{PlotExam}
\alias{PlotExam}
\title{Plot an ERGExam object.}
\usage{
PlotExam(X)
}
\arguments{
\item{X}{An ERGExam object.}
}
\value{
A ggplot2 plot of the ERGExam data.
}
\description{
This function creates a plot of an ERGExam object, including line plots of
stimulus response curves for different label categories.
}
\examples{
# Example usage:
exam_plot <- PlotExam(erg_exam)
print(exam_plot)

}
\seealso{
\linkS4class{ERGExam} \link[ggplot2:ggplot]{ggplot2:ggplot}
}
