% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateObject.R
\name{UpdateERGExam}
\alias{UpdateERGExam}
\title{Update an old (v. 0.7 / 0.8) ERGExam object}
\usage{
UpdateERGExam(X)
}
\arguments{
\item{X}{An \linkS4class{ERGExam}, \linkS4class{ERGMeasurements}, or \linkS4class{ERGElectrode} object.}
}
\value{
An updated  \linkS4class{ERGExam} object.
}
\description{
This method can be used to update an \linkS4class{ERGExam} object created in older versions of ERGtools2.
}
