% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateMeasurements.R
\name{UpdateMeasurements}
\alias{UpdateMeasurements}
\alias{[<-,ERGExam-method}
\title{UpdateMeasurements Method for ERGExam Class}
\usage{
UpdateMeasurements(X, values)

\S4method{[}{ERGExam}(x, Marker.Name, Step, Eye, Channel, Result, Time) <- value
}
\arguments{
\item{X}{A \linkS4class{ERGExam}}

\item{values}{A data.frame containing new measurement values with columns 'Recording', 'Name', 'Voltage', and 'Time'.}

\item{Step, Eye, Channel}{Vector of values for Steps, Eyes, and Channels to subset}

\item{Time}{Numeric vector of length 2 representing the time range for data extraction.
Default is the entire time range (i.e., keep all data).}
}
\value{
An updated \link[=ERGExam]{ERGExam} object with the measurements replaced or added based on the data.frame provided by the \code{values} attribute.
}
\description{
This method updates the measurements slot in an \link[=ERGExam]{ERGExam} object with new data.
}
\details{
This method is used to update the measurements of an \link[=ERGExam]{ERGExam} object with new data. The method first checks that the 'values' data.frame contains the required columns. It then merges the existing measurements with the new data based on the 'Recording' and 'Name' columns. If a matching entry is found, the 'Voltage' and 'Time' values in the 'ERGExam' object are updated with the corresponding values from 'values'. If no matching entry is found, a new entry is added to the measurements.
}
\section{Functions}{
\itemize{
\item \code{UpdateMeasurements()}: Flexibly updates the Measurements slot of an \link[=ERGExam]{ERGExam} object. Replaces marker times and values where already existent and adds them where not.

\item \code{`[`(ERGExam) <- value}: Update parts of the Measurements slot in an ERGExam object: Sets a new time for a given marker, the corresponding value is extracted automatically from the trace.

}}
